Main functions
I have started to develop the tree grow algo but I faced a couple of problems. So I will post it when I finish it.
But I think it would be smart to leave the help functions  separate from the main .

note that get_data is directed toward my computer so you probably need to change the loc in you machine.

Helper func:
I tried to write some notes near the funcs but I think they should be quite clear in general